requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
altair @ file:///home/conda/feedstock_root/build_artifacts/altair-split_1734244716962/work
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1737819173731/work
blinker @ file:///home/conda/feedstock_root/build_artifacts/blinker_1731096409132/work
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1725267488082/work
cachetools @ file:///home/conda/feedstock_root/build_artifacts/cachetools_1740094013202/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1739515848642/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1725571112467/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1735929714516/work
click @ file:///home/conda/feedstock_root/build_artifacts/click_1734858813237/work
cryptography==44.0.1
distro==1.9.0
exceptiongroup==1.2.2
frozenlist==1.5.0
gitdb @ file:///home/conda/feedstock_root/build_artifacts/gitdb_1735887193964/work
GitPython @ file:///home/conda/feedstock_root/build_artifacts/gitpython_1735929639977/work
h11==0.14.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore==1.0.7
httpx==0.28.1
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1737420181517/work
importlib_resources @ file:///home/conda/feedstock_root/build_artifacts/importlib_resources_1736252299705/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1734823942230/work
jiter==0.8.2
jsonschema @ file:///home/conda/feedstock_root/build_artifacts/jsonschema_1733472696581/work
jsonschema-specifications @ file:///tmp/tmpk0f344m9/src
jwt==1.3.1
livekit==0.20.2
livekit-agents==0.12.15
livekit-api==0.8.2
livekit-protocol==0.9.0
markdown-it-py @ file:///home/conda/feedstock_root/build_artifacts/markdown-it-py_1733250460757/work
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1733219680183/work
mdurl @ file:///home/conda/feedstock_root/build_artifacts/mdurl_1733255585584/work
multidict==6.1.0
narwhals @ file:///home/conda/feedstock_root/build_artifacts/narwhals_1739851660861/work
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1732314280888/work/dist/numpy-2.0.2-cp39-cp39-linux_x86_64.whl#sha256=62d98eb3da9f13e6b227c430d01026b7427f341b3fdcb838430f2a9e520417b1
openai==1.64.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
pandas @ file:///home/conda/feedstock_root/build_artifacts/pandas_1736810577256/work
pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1735929703139/work
pkgutil_resolve_name @ file:///home/conda/feedstock_root/build_artifacts/pkgutil-resolve-name_1733344503739/work
propcache==0.3.0
protobuf @ file:///home/conda/feedstock_root/build_artifacts/protobuf_1731365422090/work/bazel-bin/python/dist/protobuf-5.28.3-cp39-abi3-linux_x86_64.whl#sha256=4b7eaa51273fd1ef995da09f1c9c0b12a9bc91c7f75819403907369240ccacc8
psutil==5.9.8
pyarrow==19.0.1
PyAudio==0.2.14
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic==2.10.6
pydantic_core==2.27.2
pydeck @ file:///home/conda/feedstock_root/build_artifacts/pydeck_1738346561094/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1736243443484/work
PyJWT==2.10.1
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
pyttsx3==2.98
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1706886791323/work
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1737454647378/work
referencing @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_referencing_1737836872/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1733217035951/work
rich @ file:///home/conda/feedstock_root/build_artifacts/rich_1733342254348/work/dist
rpds-py @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_rpds-py_1740153283/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
smmap @ file:///home/conda/feedstock_root/build_artifacts/smmap_1739781697784/work
sniffio==1.3.1
SpeechRecognition==3.14.1
streamlit @ file:///home/conda/feedstock_root/build_artifacts/streamlit_1740115121145/work
tenacity @ file:///home/conda/feedstock_root/build_artifacts/tenacity_1733649050774/work
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1734091811753/work
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1732615921868/work
tqdm==4.67.1
types-protobuf==4.25.0.20240417
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1733188668063/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1737541069190/work
tzlocal @ file:///home/conda/feedstock_root/build_artifacts/tzlocal_1739472000838/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1734859416348/work
watchdog @ file:///home/conda/feedstock_root/build_artifacts/watchdog_1730492870473/work
watchfiles==0.24.0
yarl==1.18.3
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
zstandard @ file:///home/conda/feedstock_root/build_artifacts/zstandard_1667296091122/work